BurgerMenu and HomePage fixed
