+++ src/app/HomePage/Squad/page.js
@@ -1,8 +1,8 @@
 export default function LeaderBoard() {
   return (
-    <div className="min-h-screen flex flex-col items-center justify-center bg-gray-100">
-      <main className="flex flex-col items-center justify-center p-6">
-        <h1 className="text-4xl font-extrabold text-gray-900 mb-4">Welcome to squad page</h1>
+    <div className="h-screen flex flex-col items-center justify-center bg-gray-100 w-full">
+      <main className="flex flex-col items-center justify-center">
+        <h1 className="text-4xl font-extrabold text-gray-900">Welcome to squad page</h1>
         <p className="text-lg text-gray-700">This is squad page</p>
       </main>
     </div>
