+++ src/app/HomePage/layout.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import HomeHeader from '../components/HomePageHeader';
+import BurgerMenu from '../components/BurgerMenu';
+
+const HomeLayout = ({children}) => {
+  return (
+    <div>
+       
+       <HomeHeader /> 
+      <div className='flex'>
+        <BurgerMenu />{children}</div>
+      
+      
+    </div>
+  );
+}
+
+export default HomeLayout;
