+++ src/app/HomePage/page.js
@@ -1,53 +1,88 @@
+'use client'
+import { useState } from 'react'
 import Head from 'next/head'
 
 export default function HomePage() {
+  const [posts, setPosts] = useState([
+    { id: 1, content: 'This is the first post!', likes: 0, comments: [] },
+    { id: 2, content: 'This is the second post!', likes: 0, comments: [] },
+  ])
+
+  const handleLike = (id) => {
+    setPosts(posts.map((post) => (post.id === id ? { ...post, likes: post.likes + 1 } : post)))
+  }
+
+  const handleComment = (id, comment) => {
+    setPosts(
+      posts.map((post) => (post.id === id ? { ...post, comments: [...post.comments, comment] } : post))
+    )
+  }
+
   return (
-    <div className="min-h-screen bg-gradient-to-r from-blue-900 to-blue-200">
-      {/* Enhanced Navbar */}
-      <header className="bg-gradient-to-r from-black to-blue-800 shadow-lg">
-        <nav className="mx-auto flex items-center justify-between p-8">
-          <div className="text-white text-3xl font-bold">MyLobby</div>
-          <div className="flex-grow flex justify-center space-x-12">
-            <a
-              href="/HomePage/LeaderBoard"
-              className="text-white text-xl font-semibold py-3 px-10 rounded-lg bg-black hover:bg-blue-900 transition duration-300 transform hover:scale-105"
-            >
-              Leaderboards
-            </a>
-            <a
-              href="/HomePage/Lobby"
-              className="text-white text-xl font-semibold py-3 px-10 rounded-lg bg-black shadow-lg hover:bg-blue-900 border-2 border-blue-800 transform scale-105"
-            >
-              Lobby
-            </a>
-            <a
-              href="/HomePage/Squad"
-              className="text-white text-xl font-semibold py-3 px-10 rounded-lg bg-black hover:bg-blue-900 transition duration-300 transform hover:scale-105"
-            >
-              Squad
-            </a>
-          </div>
-        </nav>
-      </header>
+    <div className="min-h-screen flex w-full">
+      <Head>
+        <title>Lobby</title>
+      </Head>
+
 
       {/* Main Content */}
-      <main className="flex flex-col items-center justify-center min-h-screen p-6 text-center">
-        <h1 className="text-6xl font-extrabold text-gray-900 mb-6">Welcome to MyApp</h1>
-        <p className="text-lg text-gray-700 mb-8">Explore the features and pages of our amazing app!</p>
-        <button
-          href="HomePage/Lobby"
-          className="bg-blue-600 hover:bg-blue-800 text-white font-bold py-3 px-6 rounded-lg transition duration-300 transform hover:scale-105"
-        >
-          Get Started
-        </button>
-      </main>
-
-      {/* Footer */}
-      <footer className="bg-gray-800 text-white py-4">
-        <div className="container mx-auto text-center">
-          <p>&copy; 2024 MyApp. All rights reserved.</p>
-        </div>
-      </footer>
+      <div className="w-3/4 bg-gray-100 p-10">
+        <h1 className="text-5xl font-extrabold text-gray-900 mb-10">Posts</h1>
+        {posts.map((post) => (
+          <div
+            key={post.id}
+            className="bg-white p-8 rounded-2xl shadow-xl mb-10 transform transition duration-500 hover:scale-105"
+          >
+            <p className="text-xl mb-4 text-black">{post.content}</p>
+            <div className="flex items-center mt-4">
+              <button
+                className="bg-blue-600 text-white px-6 py-3 rounded-full hover:bg-blue-700 transition duration-300 transform hover:scale-110"
+                onClick={() => handleLike(post.id)}
+              >
+                Like ({post.likes})
+              </button>
+              <CommentForm postId={post.id} handleComment={handleComment} />
+            </div>
+            <div className="mt-4 space-y-2 text-black">
+              {post.comments.map((comment, index) => (
+                <p key={index} className="border-t pt-2">
+                  {comment}
+                </p>
+              ))}
+            </div>
+          </div>
+        ))}
+      </div>
     </div>
+  )
+}
+
+function CommentForm({ postId, handleComment }) {
+  const [comment, setComment] = useState('')
+
+  const handleSubmit = (e) => {
+    e.preventDefault()
+    if (comment.trim()) {
+      handleComment(postId, comment)
+      setComment('')
+    }
+  }
+
+  return (
+    <form onSubmit={handleSubmit} className="ml-4 flex items-center">
+      <input
+        type="text"
+        className="border rounded-full px-4 py-2 mr-2 flex-1 text-black"
+        placeholder="Add a comment..."
+        value={comment}
+        onChange={(e) => setComment(e.target.value)}
+      />
+      <button
+        type="submit"
+        className="bg-green-500 text-white px-6 py-3 rounded-full hover:bg-green-600 transition duration-300 transform hover:scale-110"
+      >
+        Comment
+      </button>
+    </form>
   )
 }
