+++ src/app/components/Header.js
@@ -5,6 +5,7 @@ const header = () => {
     <>
       <div className="lg:hidden block absolute top-16 w-full left-0 right-0 bg-slate-900  ">
         <ul className=" text-center text-xl p-20">
+        
           <Link to="LeaderBoard">
             <li className="my-4 py-4 border-b border-slate-800 hover:bg-slate-800 hover:rounded"> Home </li>
           </Link>
@@ -37,7 +38,7 @@ const header = () => {
           <div className="flex-10">
             <ul className="flex gap-8 -mr-12 text-[18px]">
               <Link href="/HomePage">
-                <li className=" hover:bg-cyan-800 hover:rounded"> Home </li>
+                <li className=" hover:bg-cyan-800 hover:rounded"> TempHomePage </li>
               </Link>
 
               <Link href="/login">
