+++ src/app/layout.js
@@ -14,9 +14,6 @@ export default function RootLayout({ children }) {
   return (
     <html lang="en">
       <body className="bg-main-background bg-no-repeat bg-cover">
-        <div className="bg-gradient-to-r from-teal-700 to-blue-900 shadow-lg">
-          <Header />
-        </div>
         {children}
         <Footer />
       </body>
